fix: various issues
